requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
PyJWT>=2.8.0
cryptography>=42.0.0
bcrypt>=4.1.0
pydantic[email]>=2.7.0
